added demo seed data for DB and added code for devinit DB

+++ Inventorium.Server/Model/Variation.cs
@@ -6,7 +6,7 @@
 
         public string Name { get; set; }
         public decimal Price { get; set; }
-        public int Sku { get; set; }
+        public long Sku { get; set; }
         public bool isDefault { get; set; }
 
         public ICollection<LocalVariationQuantity> LocalVariationQuantities { get; set; }
